+++ src/main.js
@@ -8,7 +8,6 @@ const tripEvents = document.querySelector('.trip-events');
 
 const LIST_COUNT = 20;
 const points = generatePoints(LIST_COUNT);
-console.log(points);
 
 const tripPresenter = new TripPresenter (tripEvents, tripMain, navigation, filter);
